feat: Handle APPDIR environment variable for AppImage paths in config
This commit modifies the configuration module to dynamically determine the application path based on the presence of the APPDIR environment variable.

When the application is run as an AppImage, APPDIR is set to the mount point of the AppImage. The APPLICATION_PATH constant in the configuration will now be set to this value, allowing for correct resolution of paths to assets, language files, and other resources within the AppImage.

When the application is run locally (outside of an AppImage), APPLICATION_PATH will default to the current working directory, maintaining compatibility with the development environment.