feat: Implement HTML mod list export with embedded icons and --no-html option
This commit introduces the functionality to export the list of installed mods to an HTML file.

Initially, icons were embedded directly into the HTML as Data URLs. To address potential file size issues, this was temporarily changed to saving icons as separate PNG files in an 'icons' folder. However, the final implementation now embeds resized and compressed icons as Data URLs to create a self-contained HTML file.

A new command-line argument, `--no-html`, has been added to allow users to disable HTML export if desired.

Additionally, this commit includes updates to the language files to support the new HTML export feature and the `--no-html` option.