Fix: Enhance Mods folder directory warning by displaying the directory name
The warning message displayed when a directory was detected within the `Mods` folder has been updated to include the name of the directory found.

This allows users to better identify the problematic directory and take necessary actions (ensuring they are using `.zip` files and not folders).

The following line of code was modified:
- `print(lang.get_translation("utils_warning_directory_in_mods_folder"))`
+ `print(f"{lang.get_translation("utils_warning_directory_in_mods_folder").format(item_name=item.name)}")`