Fix: Use command line argument for mods path
The `get_mod_path` function was updated to correctly prioritize the `modspath` argument provided via the command line. Previously, the configured mods path was always used in certain scenarios, ignoring the command line argument. This commit ensures that if the `modspath` argument is present, it will be used to determine the mods path.